Render action buttons as a prop to AccordionTree
